Introduce monitor support in solo5
Using the --mon=<path_to_scoket> we can send commands to a solo5 running
instance. For the timebeing only one support is supported "quit" which
just exits solo5. The idea is similar to the qmp, where we can send
commands to a QEMU machine over a socket.

We can communicate with  the running solo5 instance through the UNIX
domain socket which was specified in --mon option.

This feauture will be used to invoke a migration command.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>