Handle monitor signals while polling
Solo5 provides a poll hypercall, which can be used from guests to get
notified for any events, network I/O etc. If the guest just waits for an
event, then the vCPU loop is blocked in the poll hypercall. As a reuslt
the SIG_USR1 signal, used by monitor, was blocked by the ppoll system
call.

A signal from solo5_monitor hich results to a change in VM state will
interrupt the ppoll system call and then solo5 will handle the change in
VM state. Moreover 0 will be the return value of poll hypercall, since
Solo5 assures the guests that only a non negative value will be
returned.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>