Move setup of memory permissions after creating VM
Solo5 sets memory permissions to guest memory according to the prgram
header flags of the guest kernel elf. These permissions are enforced
while loading the elf binary. However, this might create problems in
migration. Some pages in guest memory might need to be overwritten, but
they might not have the correct permissions.

In order to resolve this issue, the memory permissions should be enforced
after the completion of migration.

The code which sets up the permissions is moved after setting up the VM
and before the vCPU loop starts.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>