Introduce save VM monitor command
The command savevm is the first part of the migration support. Using
this command solo5 will save the current state of the guest VM in a file
specified during the command. This file will be later used to load the
VM state to a new instance.

The syntax of command is:
savevm <file_to_save_VM>

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>