Scheduler: Create frontend thread and socket
Create a thread which will wait for new binaries to be added in the
scheduler. In order to add new binaries, the full path of the binary
has to be sent to /tmp/front.sock.

The frontend will be a REST API which will be used to upload new
binaries, response to queries regarding the scheduler (how many nodes
are available, how many tasks are queued etc) and maybe control
the scheduler.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>