Scheduler: Switch from signals to eventfd and epoll
Signal communication between the main thread (scheduler) and the thread
which receives the new task has been replaced from a epoll amd eventfd
scheme.

There will be only one thread, the scheduler and it will uses other
threads as workers in order to communicate with the outter world.
Every new connection in the frontend socket (it will be used for new
binaries) will be handled by a new thread, which is responsible to
create the new task struct and notify the scheduler for the new task.

The communication between the worker threads and the scheduler will be
done through eventfds.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>