Scheduler: Send result of binary execution to primary
When daemon scheduler receives the new binary it will execute it and
wait for the result or for a new command from the primary scheduler. If
the executions is over then the daemon will report back to the primary
the change in state and the result of execution (success or fail).

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>