Send migration command
With the command "Task: {id} Node: {id}" from frontend the primary
scheduler will send the ip address of the new node where the task will
be migrated.

When a node receives such info it will initiate the migration proccess
in ukvm and report the result back to primary scheduler.

Moreover daemon should be able to distingusih between a failed exit or
the completion of migration. In order to do that ukvm will return the
value 7 to indicate the successful creation of a migration file.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>