Scheduler: Migrate task to a different node
When a node receives a migration command it will set up a connection
with the new node where the task will run. The initial node will send
both the binary and the migration file directly to the new node.

After receiving theses 2 files the new node will start ukvm using
"--load" option and the task will continue its execution in the new node.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>