Scheduler: Create common file for primary and daemon
There is common code between primary and daemon. That code will be
placed in a new file common.c and a new header file will be used
common.h

Moreover intorduce err_print() function to format messages for stderr.
Using this macro messages will be printed with the fuction and line of
code where the caro got called.

In addition some comments were added and the algorithm of scheduler got
decoupled from the main function, in order to be easier to change it.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>