scheduler: Add support for task priority
From now on the user can define the priority of a task.
There are 2 possible priorities:
- 0: High priority
- 1: Low priority

There are 2 different lists, one for each priority. Both lists are
combined in one when the algorithm selects an application, placing the
low priority tasks after the high priority ones. The algorithm remains
the same for the time being (FCFS).

In order to specify the priority the user should send a command like:
"New: <path_to_guest> priority: 1"
in the /tmp/front.sock UNIX domain socket.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>