Scheduler: Add eviction support
When a new task with high priority is added to the scheduler then this
task will be deployed in a node that runs a low priority task. The low
priority task will get evicted (syopped) and a new ukvm instance will
start with the high priority task. When task finishes the previous task
will resume its execution.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>