Scheduler: Update migration handling and store ukvm arguments
Daemon scheduler became more flexible, being able to control 2 ukvm
processes. Moreover daemon will have to keep track of ukvm arguments to
select the correct socket, binary when communicating with an instance.

Also this commit includes and a small change in primary scheduler, which
was needed when searching for task to migrate, because of the split of
task list in 2, low and high priority.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>