Scheduler: Let daemon now about task id
For a more stable and less error prone communication between primary and
daemon scheduler, the task is will be used as an identifier of the
respective task. Being able to execute more than one tasks in the same
time from daemon increases complexity and there might be some race
conditions in messages between primary and daemon.

For instance the primary might decide to evict a task and while sending
the command to the daemon the task to be evicted might terminate before
receiving the eviction command. Using the task id, daemon will be able to
handle this situation reporting back to primary the corresponding exit
code.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>