Support migration over eviction and fix minor bug in migration
An evicted application is now able to get migrated using the same
migration command as an usual application. This feature is needed in
case scheduler decides to migrate an application which has previously
evicted.

Moreover a minor bug in migration is fixed. In case the fpga worker
exists but there are no fpga data, the worker was not getting terminated,
since it was never waiting for a message to do so.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>